fix: 授業取得ページを作成
